Progredindo com o DropDown

+++ backend/src/main/java/com/basis/sgcproject/service/CategoriaService.java
@@ -1,6 +1,7 @@
 package com.basis.sgcproject.service;
 
 import com.basis.sgcproject.repository.CategoriaRepository;
+import com.basis.sgcproject.repository.CompetenciaRepository;
 import com.basis.sgcproject.service.dto.CategoriaDTO;
 import com.basis.sgcproject.service.mapper.CategoriaMapper;
 import lombok.AllArgsConstructor;
@@ -16,6 +17,11 @@ public class CategoriaService {
     private final CategoriaMapper categoriaMapper;
 
     private final CategoriaRepository categoriaRepository;
+
+    private final CompetenciaRepository competenciaRepository;
+
+    private final CategoriaCompetenciaService categoriaCompetenciaService;
+
 
     public List<CategoriaDTO> buscar(){
 
@@ -35,9 +41,14 @@ public class CategoriaService {
     }
 
     public void delete(Integer id){
-
+        if (!categoriaRepository.existsById(id)){
+            throw new RuntimeException("Esse ID não existe!");
+        }
+        removerCategoriaDeCompetencia(id);
         categoriaRepository.deleteById(id);
-
     }
 
+    public void removerCategoriaDeCompetencia(Integer categoriaId){
+        categoriaCompetenciaService.removerCategoriasDeCompetecias(categoriaId);
+    }
 }
